fix(*): errors get address ETH, ERC1155 tokens, rename, delete needless tests
