--- exchange-v2/test-exchangev2-doTransfer.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-truffle test ./test/v2/RaribleManagerDoTransfer.test.js ./test/contracts/tokens/TestERC20.sol ./test/contracts/tokens/TestERC721.sol ./test/contracts/tokens/TestERC1155.sol ./test/contracts/v2/RaribleTransferManagerTest.sol ./test/contracts/v2/LibOrderTest.sol
